src/bot/main.py
#!/usr/bin/env python3
"""
BrislyDeals Gaming Bot - Main Entry Point
Telegram: @BrislyGamingBot
Channel: @BrislyDealsGaming
"""

import os
import sys
import logging
from datetime import datetime
from dotenv import load_dotenv

# Carica le variabili d'ambiente
load_dotenv()

# Configurazione logging
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

class BrislyGamingBot:
    """Main bot class per gestire le operazioni"""
    
    def __init__(self):
        self.bot_token = os.getenv('TELEGRAM_BOT_TOKEN')
        self.channel_id = os.getenv('TELEGRAM_CHANNEL_ID')
        
        if not self.bot_token:
            logger.error("❌ TELEGRAM_BOT_TOKEN non trovato!")
            sys.exit(1)
            
        logger.info("✅ BrislyGaming Bot inizializzato")
        logger.info(f"📢 Canale target: {self.channel_id}")
        
    def run(self):
        """Avvia il bot"""
        logger.info(f"🚀 Bot avviato alle {datetime.now()}")
        logger.info("🔄 In attesa di implementazione scraper...")
        
        # TODO: Implementare logica principale
        # - Avviare scrapers
        # - Schedulare posting
        # - Gestire comandi
        
        print("\n" + "="*50)
        print("🎮 BRISLY GAMING BOT - READY")
        print("="*50)
        print(f"Channel: {self.channel_id}")
        print(f"Time: {datetime.now()}")
        print("Status: Waiting for implementation")
        print("="*50 + "\n")

def main():
    """Entry point principale"""
    try:
        bot = BrislyGamingBot()
        bot.run()
    except KeyboardInterrupt:
        logger.info("⏹️ Bot fermato manualmente")
    except Exception as e:
        logger.error(f"❌ Errore critico: {e}")
        sys.exit(1)

if __name__ == "__main__":
    main()